Consolidate form helpers
Removing the `require "active_admin/view_helpers/method_or_proc_helper"` in `lib/active_admin/filters.rb` hid a loading bug with `lib/active_admin/resource.rb` since some of its classes use helper methods from method_or_proc_helper (e.g. scopes, action_items, etc.) but included the module without requiring it. We now require it at the top of the resource file.

This brings together the original form_helper and fields_for file which the latter oddly didn't declare a different module but used form helper instead.

Since the filter form helper spec included a lot of tests, we'll keep this spec separate from form helper despite being in the same one.
