+++ lib/active_admin/resource.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require "active_admin/view_helpers/method_or_proc_helper"
 require "active_admin/resource/action_items"
 require "active_admin/resource/attributes"
 require "active_admin/resource/controllers"
